Nearly Final Product
Got to go to a shift at work, however nearly done for the most part. I wont be able to make the models in time liek I wanted or the UI, however the overall gameplay loop I wanted is done. There are a couple bugs like, audio and the map pointer keeps on destroying. But for the most part it is done and will complete the rest of teh requirements after my shift. I need to complete and options menu and get rid of some bugs then I should be all done